pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vivek"
version = "2.0.0"
requires-python = ">=3.11"
description = "Privacy-first collaborative AI brain design for intelligent coding assistance"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Sanket Naik", email = "sanketn@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "ollama>=0.3.0",
    "requests>=2.28.0",
    "gitpython>=3.1.0",
    "pathspec>=0.11.0",
    "watchdog>=3.0.0",
    "psutil>=5.9.0",
    "colorama>=0.4.6",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "typing-extensions>=4.5.0",
    "tiktoken>=0.5.0",
    "numpy (>=2.3.3,<3.0.0)",
    "sentence-transformers (>=5.1.1,<6.0.0)",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-click>=1.1.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
vivek = "src.vivek.cli:cli"

[project.urls]
Homepage = "https://github.com/sanketn26/vivek"
Repository = "https://github.com/sanketn26/vivek"
Documentation = "https://github.com/sanketn26/vivek#readme"
"Bug Tracker" = "https://github.com/sanketn26/vivek/issues"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["vivek*"]

[tool.setuptools.package-data]
"vivek" = ["*.md", "*.yml", "*.yaml"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "cli: CLI functionality tests",
    "llm: LLM model tests",
    "orchestrator: Orchestrator tests",
    "asyncio: Async tests"
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning"
]